Heavily cleanup pluginlib.
In particular:

1.  Move pluginlib to the top directory.  The main reason
to do this is so that our copyright checks actually work
properly here.  But this also gets rid of an unnecessary
directory.
2.  Enable linting.  This ensures that this project adheres
to our general guidelines.
