+++ include/pluginlib/impl/split.hpp
@@ -1,0 +1,55 @@
+// Copyright (c) 2017, Open Source Robotics Foundation, Inc. All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//    * Redistributions of source code must retain the above copyright
+//      notice, this list of conditions and the following disclaimer.
+//
+//    * Redistributions in binary form must reproduce the above copyright
+//      notice, this list of conditions and the following disclaimer in the
+//      documentation and/or other materials provided with the distribution.
+//
+//    * Neither the name of the Willow Garage nor the names of its
+//      contributors may be used to endorse or promote products derived from
+//      this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+#ifndef PLUGINLIB__IMPL__SPLIT_HPP_
+#define PLUGINLIB__IMPL__SPLIT_HPP_
+
+#include <regex>
+#include <string>
+#include <vector>
+
+namespace pluginlib
+{
+namespace impl
+{
+
+inline std::vector<std::string>
+split(const std::string & input, const std::string & regex)
+{
+  std::regex re(regex);
+  // the -1 will cause this to return the stuff between the matches, see the submatch argument:
+  //   http://en.cppreference.com/w/cpp/regex/regex_token_iterator/regex_token_iterator
+  std::sregex_token_iterator first(input.begin(), input.end(), re, -1);
+  std::sregex_token_iterator last;
+  return {first, last};  // vector will copy from first to last
+}
+
+}  // namespace impl
+}  // namespace pluginlib
+
+#endif  // PLUGINLIB__IMPL__SPLIT_HPP_
