--- pluginlib/test/include/test_plugins.h
@@ -1,83 +1,0 @@
-/*
- * Copyright (c) 2012, Willow Garage, Inc.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the
- *       documentation and/or other materials provided with the distribution.
- *     * Neither the name of the Willow Garage, Inc. nor the names of its
- *       contributors may be used to endorse or promote products derived from
- *       this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
- * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
- * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
- * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
- * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
- * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
- * POSSIBILITY OF SUCH DAMAGE.
- */
-
-#ifndef TEST_PLUGINS_H_
-#define TEST_PLUGINS_H_
-
-#include <cmath>
-
-#include <test_base.h>
-#include <visibility_control.hpp>
-
-namespace test_plugins
-{
-class TEST_PLUGINLIB_FIXTURE_PUBLIC Bar : public test_base::Fubar
-{
-public:
-  Bar() {}
-
-  void initialize(double foo)
-  {
-    foo_ = foo;
-  }
-
-  double result()
-  {
-    return 0.5 * foo_ * getBar();
-  }
-
-  double getBar()
-  {
-    return sqrt((foo_ * foo_) - ((foo_ / 2) * (foo_ / 2)));
-  }
-
-private:
-  double foo_;
-};
-
-class TEST_PLUGINLIB_FIXTURE_PUBLIC Foo : public test_base::Fubar
-{
-public:
-  Foo() {}
-
-  void initialize(double foo)
-  {
-    foo_ = foo;
-  }
-
-  double result()
-  {
-    return foo_ * foo_;
-  }
-
-private:
-  double foo_;
-};
-}  // namespace test_plugins
-#endif  // TEST_PLUGINS_H_
