+++ test/test_plugins.xml
